requirements.txt
fastapi
httpx
python-dotenv
motor
uvicorn
